Added tcp and udp classes, have not tested. Added service folder and added a background service to it, this will listen for any connections. Current thought process is:
A udp and tcp listener each listen for connections from radios.

If a connection is received it triggers a variable inside the tcp/udp class called EarsFull, this is a bool that is true if a message is currently being transmitted.

We do not want to cut this message off so for as long as a message is being transmitted, the receiver's "Ears" are full and it cannot listen to other incoming connections.

Therefore this connection is added to a list of IListeners, an interface of both udp and tcp listeners.

A new instance of the listener is then created outside of the list to continue listening for connections.

Just need to add and debug tcp class and udp class too tbf, then hopefully debug the background service, then ensure class can have messages parsed to the list which should display them on a live feed kinda, then do more stuff.